Added diagnostics (cpp+py) (#16)
### Changelog
Added diagnostics tutorial code, both in C++ and Python

### Docs
N/A

### Description

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

New packages for simple diagnostics publisher, in C++ and Python.

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->

---------

Co-authored-by: Hans-Joachim Krauch <achim-k@users.noreply.github.com>