Add foxglvoe so_100 viewer tutorial files (#29)
### Changelog
This PR adds tutorial files for visualizing SO-ARM100 with LeRobot and
Foxglove.

### Docs

None

### Description

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->
No script to visualize SO-100 with Foxglove
</td><td>

<!--after content goes here-->
There is a script to visualize SO-100 with Foxglove
</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->

---------

Co-authored-by: banisadr <banisadr@users.noreply.github.com>