Refactor tutorial repo to support automatic README generation (#33)
This PR updates the structure of the entire tutorial repository, with
the goal of making it possible to run all the tutorials without
following the blog posts.

Thanks to the Python script in the .utils directory, we can autogenerate
the main README file using frontmatter defined for each tutorial.

Once this is merged, I'll create a new PR for a github action to
automatically regen README on PR.

Changes introduced in this PR:
* Created a new set of top-level directories to try to categorize
tutorials a little bit in the following categories:
  - datasets
  - foxglove_sdk
  - integrations (currently holds ros1 and ros2 directories)
  - jupyter_notebooks
* Created a frontmatter structure for README files and applied it to
every existing tutorial
* Created a template for the top-level README and a script to regenerate
it
* Added a CONTRIBUTING.md file describing the workflow to contribute a
tutorial

---------

Co-authored-by: Bryan Fox <39674+bryfox@users.noreply.github.com>